% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCountPhe.R
\name{createCountPhe}
\alias{createCountPhe}
\title{createCountPhe}
\usage{
createCountPhe()
}
\value{
a list
}
\description{
This function would generate a list including a countData and colData as input data for testing
}
\examples{
input = createCountPhe()
}
