% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEseqObj.R
\name{DEseqObj}
\alias{DEseqObj}
\title{DEseqObj}
\usage{
DEseqObj(
  count_data,
  col_data,
  design_names = "condition+type",
  group_name = "condition",
  ref_level = "control"
)
}
\arguments{
\item{count_data}{a matrix with row name}

\item{col_data}{a dataframe include group information}

\item{design_names}{designs with multiple variables, e.g., group + condition, and designs with interactions, e.g., genotype + treatment + genotype:treatment.}

\item{group_name}{the name of col_data to split case and control group}

\item{ref_level}{a character which is the reference group}
}
\value{
a list
}
\description{
This function would generate a object of DESeq2 class
}
\examples{
\dontrun{
input = createCountPhe()
dds = DEseqObj(input[[1]],input[[2]],ref_level="untreated")
}
}
