% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeqObjPCA.R
\name{DESeqObjPCA}
\alias{DESeqObjPCA}
\title{DESeqObjPCA}
\usage{
DESeqObjPCA(
  dds,
  group_name = "condition",
  label = FALSE,
  outputDir = ".",
  base_name = NULL
)
}
\arguments{
\item{dds}{a matrix with row name}

\item{group_name}{the number of colData to split case and control group}

\item{label}{A logical value to plot label of sample}

\item{outputDir}{the path of output}

\item{base_name}{NULL or character, the prefix of output. When value is NULL, the figures are not saved.}
}
\description{
This function would plot PCA based on DESeqDataSet object and save as pdf and png format
}
\examples{
\dontrun{
input = createCountPhe()
dds = DEseqObj(input[[1]],input[[2]],ref_level="untreated")
DESeqObjPCA(dds)
}
}
