% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeqRes.R
\name{DESeqRes}
\alias{DESeqRes}
\title{DESeqRes}
\usage{
DESeqRes(dds, foldChange = 1, adjPvalue = 0.05)
}
\arguments{
\item{dds}{a matrix with row name}

\item{foldChange}{a dataframe include group information}

\item{adjPvalue}{the number of colData to split case and control group}
}
\value{
a dataframe
}
\description{
This function would get the results based on DESeqDataSet object
}
\examples{
\dontrun{
input = createCountPhe()
dds = DEseqObj(input[[1]],input[[2]],refLevel="untreated")
res = DESeqRes(dds)
}
}
