% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeqResVolcano.R
\name{DESeqResVolcano}
\alias{DESeqResVolcano}
\title{DESeqResVolcano}
\usage{
DESeqResVolcano(
  DEGres,
  base_name = NULL,
  outputDir = ".",
  foldChange = 1,
  adjPvalue = 0.05
)
}
\arguments{
\item{DEGres}{a dataframe which would be generated by DESeqRes function}

\item{base_name}{NULL or character, the prefix of output. When value is NULL, the figures are not saved.}

\item{outputDir}{the path of output}

\item{foldChange}{a dataframe include group information}

\item{adjPvalue}{the number of colData to split case and control group}
}
\description{
This function would plot volcano based on DESeqRes output and save as pdf and png format
}
\examples{
\dontrun{
input = createCountPhe()
dds = DEseqObj(input[[1]],input[[2]],refLevel="untreated")
res = DESeqRes(dds)
DESeqResVolcano(res)
}
}
