% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{GO}
\alias{GO}
\title{GO}
\usage{
GO(
  ids,
  type = "ENTREZID",
  db = org.Hs.eg.db,
  outputDir = ".",
  base_name = NULL,
  ...
)
}
\arguments{
\item{ids}{A character vector of gene id}

\item{type}{The type of ids which must be one of idType(OrgDb = db) shows}

\item{db}{The database, such as org.Hs.eg.db , org.Mm.eg.db and so on.}

\item{outputDir}{The path of output}

\item{base_name}{The prefix name of outputs would be saved}

\item{...}{Aavailable arguments to be passed to enrichGO.}
}
\description{
This function would plot volcano based on DESeqDEGres output and save as pdf and png format
}
