% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAgo.R
\name{GSEAgo}
\alias{GSEAgo}
\title{GSEAgo}
\usage{
GSEAgo(geneChangeList, db = "org.Hs.eg.db", outputDir = ".", baseName = "DEG")
}
\arguments{
\item{geneChangeList}{a numeric vector of the fold change of genes. It must have name with ENTREZID.}

\item{db}{the database, such as org.Hs.eg.db , org.Mm.eg.db and so on.}

\item{outputDir}{the path of output}

\item{baseName}{the prefix name of outputs would be saved}
}
\description{
This function would plot volcano based on DESeqDEGres output and save as pdf and png format
}
