% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAkegg.R
\name{GSEAkegg}
\alias{GSEAkegg}
\title{GSEAkegg}
\usage{
GSEAkegg(geneChangeList, organism = "hsa", outputDir = ".", baseName = "DEG")
}
\arguments{
\item{geneChangeList}{a numeric vector of the fold change of genes. It must have name with ENTREZID.}

\item{organism}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html', such as hsa for human, mmu for mouse.}

\item{outputDir}{the path of output}

\item{baseName}{the prefix name of outputs would be saved}
}
\description{
This function would plot volcano based on DESeqDEGres output and save as pdf and png format
}
