% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{KEGG}
\alias{KEGG}
\title{KEGG}
\usage{
KEGG(
  ids,
  type = "ENTREZID",
  organism = "hsa",
  db = org.Hs.eg.db,
  outputDir = ".",
  base_name = NULL,
  ...
)
}
\arguments{
\item{ids}{a character vector of gene id}

\item{type}{the type of ids which must be one of idType(OrgDb = db) shows}

\item{organism}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html', such as hsa for human, mmu for mouse.}

\item{db}{the annotation database of organism, such as org.Hs.eg.db for human}

\item{outputDir}{the path of output}

\item{base_name}{the prefix name of outputs would be saved}

\item{...}{Aavailable arguments to be passed to enrichKEGG}
}
\description{
This function would plot volcano based on DESeqDEGres output and save as pdf and png format
}
