% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqFlow.R
\name{RNAseqFlow}
\alias{RNAseqFlow}
\title{DESeqRes}
\usage{
RNAseqFlow(
  geneCountFile,
  pheFile,
  outputDir = ".",
  groupNum = 1,
  foldChange = 1,
  adjPvalue = 0.05,
  db = "org.Hs.eg.db",
  organism = "hsa"
)
}
\arguments{
\item{geneCountFile}{a matrix with row name}

\item{pheFile}{a dataframe include group information}

\item{outputDir}{the number of colData to split case and control group}

\item{groupNum}{the number of colData to split case and control group}

\item{foldChange}{the significant threshold for log2 fold change}

\item{adjPvalue}{the significant threshold for adjust P-value}

\item{db}{the database, such as org.Hs.eg.db , org.Mm.eg.db and so on.}

\item{organism}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html', such as hsa for human, mmu for mouse.}
}
\value{
a dataframe
}
\description{
This function would get the results based on DESeqDataSet object
}
