% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEseqObj.R
\name{DEseqObj}
\alias{DEseqObj}
\title{DEseqObj}
\usage{
DEseqObj(countData, colData, groupNum = 1, refLevel = "control")
}
\arguments{
\item{countData}{a matrix with row name}

\item{colData}{a dataframe include group information}

\item{groupNum}{the number of colData to split case and control group}

\item{refLevel}{a character which is the reference group}
}
\value{
a list
}
\description{
This function would generate a object of DESeq2 class
}
\examples{
\dontrun{
input = createCountPhe()
dds = DEseqObj(input[[1]],input[[2]],refLevel="untreated")
}
}
