% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeqObjPCA.R
\name{DESeqObjPCA}
\alias{DESeqObjPCA}
\title{DESeqObjPCA}
\usage{
DESeqObjPCA(dds, groupNum = 1, outputDir = ".")
}
\arguments{
\item{dds}{a matrix with row name}

\item{groupNum}{the number of colData to split case and control group}

\item{outputDir}{the path of output}
}
\description{
This function would plot PCA based on DESeqDataSet object and save as pdf and png format
}
\examples{
\dontrun{
input = createCountPhe()
dds = DEseqObj(input[[1]],input[[2]],refLevel="untreated")
DESeqObjPCA(dds)
}
}
