% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{GO}
\alias{GO}
\title{GO}
\usage{
GO(
  ids,
  type = "SYMBOL",
  db = "org.Hs.eg.db",
  outputDir = ".",
  baseName = "DEG"
)
}
\arguments{
\item{ids}{a character vector of gene id}

\item{type}{the type of ids which must be one of idType(OrgDb = db) shows}

\item{db}{the database, such as org.Hs.eg.db , org.Mm.eg.db and so on.}

\item{outputDir}{the path of output}

\item{baseName}{the prefix name of outputs would be saved}
}
\description{
This function would plot volcano based on DESeqDEGres output and save as pdf and png format
}
