% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEAgo}
\alias{GSEAgo}
\title{GSEAgo}
\usage{
GSEAgo(geneChangeList, db = org.Hs.eg.db)
}
\arguments{
\item{geneChangeList}{a numeric vector of the fold change of genes. It must have name with ENTREZID.}

\item{db}{the database, such as org.Hs.eg.db , org.Mm.eg.db and so on.}
}
\description{
This function would plot volcano based on DESeqDEGres output and save as pdf and png format
}
\examples{
\dontrun{
data(geneList, package="DOSE")
ego = GSEAgo(geneList)
}
}
