% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{KEGG}
\alias{KEGG}
\title{GO}
\usage{
KEGG(ids, type = "SYMBOL", organism = "hsa", outputDir = ".", baseName = "DEG")
}
\arguments{
\item{ids}{a character vector of gene id}

\item{type}{the type of ids which must be one of idType(OrgDb = db) shows}

\item{organism}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html', such as hsa for human, mmu for mouse.}

\item{outputDir}{the path of output}

\item{baseName}{the prefix name of outputs would be saved}
}
\description{
This function would plot volcano based on DESeqDEGres output and save as pdf and png format
}
